+++ packages/core/src/integrations/index.ts
@@ -1,3 +1,2 @@
-export { Dedupe } from './dedupe';
 export { FunctionToString } from './functiontostring';
 export { InboundFilters } from './inboundfilters';