+++ packages/integrations/src/index.ts
@@ -1,2 +1,0 @@
-// We do not export anything by default since we always want to deep import
-// like: import { Vue } from '@sentry/integrations/vue';