+++ packages/node/src/sdk.ts
@@ -7,7 +7,6 @@ import { Console, Http, LinkedErrors, OnUncaughtException, OnUnhandledRejection 
 
 export const defaultIntegrations = [
   // Common
-  new CoreIntegrations.Dedupe(),
   new CoreIntegrations.InboundFilters(),
   new CoreIntegrations.FunctionToString(),
   // Native Wrappers